+++ browser-extension/api-server/routes/feedback.js
@@ -1,0 +1,13 @@
+//console.log(`Environment : ${process.env.NODE_ENV}`);
+
+const express = require("express");
+const router = express.Router();
+
+
+const { addFeedback } = require("../controllers/feedback");
+
+
+router.post("/feedback", addFeedback);
+
+
+module.exports = router  ;
