+++ browser-extension/api-server/routes/predict.js
@@ -1,0 +1,14 @@
+//console.log(`Environment : ${process.env.NODE_ENV}`);
+
+const express = require("express");
+const router = express.Router();
+
+
+const {predict} = require("../controllers/predict") ;
+
+
+router.post("/predict", predict);
+
+
+
+module.exports = router ;
